Fix spatial angle header ownership

+++ include/game/math/spatial_math.h
@@ -4,7 +4,7 @@
 #include "common.h"
 
 s32 isPositionNearCurrentRaceViewportCamera(Vec3i *position);
-s32 calculateFixedAngleFromDeltaXZ(s32 deltaX, s32 deltaZ);
-s32 calculateFixedAngleBetweenXZPoints(s32 fromX, s32 fromZ, s32 toX, s32 toZ);
+s16 calculateFixedAngleFromDeltaXZ(s32 deltaX, s32 deltaZ);
+s16 calculateFixedAngleBetweenXZPoints(s32 fromX, s32 fromZ, s32 toX, s32 toZ);
 
 #endif
